prog, new json parser

+++ bhbb_dl/include/main.h
@@ -9,9 +9,9 @@ SCE_CDECL_BEGIN
 SceInt32 module_start(SceSize args, void *argp);
 SceInt32 module_stop(SceSize args, void *argp);
 
-typedef void*(*curl_malloc)(unsigned int size);
+typedef void*(*curl_malloc)(size_t size);
 typedef void(*curl_f)(void *ptr);
-typedef void*(*curl_realloc)(void *ptr, unsigned int new_size);
+typedef void*(*curl_realloc)(void *ptr, size_t new_size);
 
 int curl_global_memmanager_set_np(curl_malloc allocate, curl_f deallocate, curl_realloc reallocate);
 int zipInitPsp2(void);
